refactor sitemeta package
