+++ sitemeta/site_meta.go
@@ -1,15 +1,11 @@
 package sitemeta
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"strings"
 
 	iconv "github.com/djimenez/iconv-go"
-
-	content "github.com/hirakiuc/site-meta-go/sitemeta/content"
 )
 
 const (
@@ -17,16 +13,9 @@ const (
 	DefaultEncoding string = "UTF-8"
 )
 
-//nolint
-var CantParseErr error = errors.New("this site can't parse SiteMeta")
-
 // SiteMeta describe meta data of the website, like ogp, TwitterCard.
 type SiteMeta struct {
 	Attrs map[string]string
-}
-
-func newSiteMeta() SiteMeta {
-	return SiteMeta{Attrs: map[string]string{}}
 }
 
 // String return a description about this instance.
@@ -37,8 +26,17 @@ func (meta *SiteMeta) String() string {
 		str := fmt.Sprintf("%s - %s", key, value)
 		attrs = append(attrs, str)
 	}
-
 	return strings.Join(attrs, "\n")
+}
+
+func (meta *SiteMeta) AddMeta(key string, val string) bool {
+	if !strings.HasPrefix(key, "twitter:") && !strings.HasPrefix(key, "og:") {
+		return false
+	}
+
+	meta.Attrs[key] = val
+
+	return true
 }
 
 func (meta *SiteMeta) ToJSON() (string, error) {
@@ -50,10 +48,14 @@ func (meta *SiteMeta) ToJSON() (string, error) {
 	return string(bytes), nil
 }
 
+func (meta *SiteMeta) IsEmpty() bool {
+	return len(meta.Attrs) == 0
+}
+
 // IsValid validate that this instance keeps valid value, or not.
 func (meta *SiteMeta) IsValid() bool {
-	if len(meta.Attrs) == 0 {
-		return false
+	if meta.IsEmpty() {
+		return true
 	}
 
 	for key := range meta.Attrs {
@@ -97,27 +99,4 @@ func (meta *SiteMeta) convertEncoding(toEncoding string) error {
 	meta.Attrs = result
 
 	return nil
-}
-
-// Parse return SiteMeta instance if the url content has meta tags about twitter card or ogp.
-func Parse(c context.Context, url string) (*SiteMeta, error) {
-	data := newSiteMeta()
-
-	html, err := content.FetchHTMLContent(c, url)
-	if err != nil {
-		return nil, err
-	}
-
-	data.Attrs = html.MetaAttrs()
-
-	err = data.convertEncoding(html.ContentEncoding)
-	if err != nil {
-		return nil, err
-	}
-
-	if !data.IsValid() {
-		return nil, fmt.Errorf("%w", CantParseErr)
-	}
-
-	return &data, nil
 }
